+++ Src/MoneyFox.Uwp/MoneyFox.Uwp.csproj
@@ -186,14 +186,14 @@
     <Compile Include="Src\WindowsAppInformation.cs" />
     <Compile Include="Src\WindowsFileStore.cs" />
     <Compile Include="ViewModelLocator.cs" />
-    <Compile Include="ViewModels\AbstractCategoryListViewModel.cs" />
-    <Compile Include="ViewModels\AccountListViewActionViewModel.cs" />
-    <Compile Include="ViewModels\AccountListViewModel.cs" />
-    <Compile Include="ViewModels\AddAccountViewModel.cs" />
-    <Compile Include="ViewModels\AddCategoryViewModel.cs" />
+    <Compile Include="ViewModels\Categories\AbstractCategoryListViewModel.cs" />
+    <Compile Include="ViewModels\Accounts\AccountListViewActionViewModel.cs" />
+    <Compile Include="ViewModels\Accounts\AccountListViewModel.cs" />
+    <Compile Include="ViewModels\Accounts\AddAccountViewModel.cs" />
+    <Compile Include="ViewModels\Categories\AddCategoryViewModel.cs" />
     <Compile Include="ViewModels\Payments\AddPaymentViewModel.cs" />
     <Compile Include="ViewModels\BalanceViewModel.cs" />
-    <Compile Include="ViewModels\CategoryListViewModel.cs" />
+    <Compile Include="ViewModels\Categories\CategoryListViewModel.cs" />
     <Compile Include="ViewModels\DesignTime\DesignTimeAccountListViewActionViewModel.cs" />
     <Compile Include="ViewModels\DesignTime\DesignTimeAccountListViewModel.cs" />
     <Compile Include="ViewModels\DesignTime\DesignTimeBalanceViewViewModel.cs" />
@@ -212,27 +212,27 @@
     <Compile Include="ViewModels\RecurringPaymentViewModel.cs" />
     <Compile Include="ViewModels\Statistic\StatisticCategorySummary\DesignTimeStatisticCategorySummaryViewModel.cs" />
     <Compile Include="ViewModels\DesignTime\DesignTimeStatisticSelectorViewModel.cs" />
-    <Compile Include="ViewModels\EditAccountViewModel.cs" />
-    <Compile Include="ViewModels\EditCategoryViewModel.cs" />
+    <Compile Include="ViewModels\Accounts\EditAccountViewModel.cs" />
+    <Compile Include="ViewModels\Categories\EditCategoryViewModel.cs" />
     <Compile Include="ViewModels\Payments\EditPaymentViewModel.cs" />
-    <Compile Include="ViewModels\ICategoryListViewModel.cs" />
-    <Compile Include="ViewModels\IModifyCategoryViewModel.cs" />
+    <Compile Include="ViewModels\Categories\ICategoryListViewModel.cs" />
+    <Compile Include="ViewModels\Categories\IModifyCategoryViewModel.cs" />
     <Compile Include="ViewModels\Payments\IModifyPaymentViewModel.cs" />
     <Compile Include="ViewModels\IncomeExpenseBalanceViewModel.cs" />
     <Compile Include="ViewModels\Interfaces\IAccountListViewActionViewModel.cs" />
     <Compile Include="ViewModels\Interfaces\IAccountListViewModel.cs" />
     <Compile Include="ViewModels\Interfaces\IBalanceViewModel.cs" />
     <Compile Include="ViewModels\Interfaces\IPaymentListViewActionViewModel.cs" />
-    <Compile Include="ViewModels\ISelectCategoryListViewModel.cs" />
+    <Compile Include="ViewModels\Categories\ISelectCategoryListViewModel.cs" />
     <Compile Include="ViewModels\ISelectDateRangeDialogViewModel.cs" />
     <Compile Include="ViewModels\ISelectFilterDialogViewModel.cs" />
-    <Compile Include="ViewModels\ModifyAccountViewModel.cs" />
-    <Compile Include="ViewModels\ModifyCategoryViewModel.cs" />
+    <Compile Include="ViewModels\Accounts\ModifyAccountViewModel.cs" />
+    <Compile Include="ViewModels\Categories\ModifyCategoryViewModel.cs" />
     <Compile Include="ViewModels\Payments\ModifyPaymentViewModel.cs" />
     <Compile Include="ViewModels\Payments\PaymentListBalanceViewModel.cs" />
     <Compile Include="ViewModels\Payments\PaymentListViewActionViewModel.cs" />
     <Compile Include="ViewModels\Payments\PaymentListViewModel.cs" />
-    <Compile Include="ViewModels\SelectCategoryListViewModel.cs" />
+    <Compile Include="ViewModels\Categories\SelectCategoryListViewModel.cs" />
     <Compile Include="ViewModels\SelectDateRangeDialogViewModel.cs" />
     <Compile Include="ViewModels\SelectFilterDialogViewModel.cs" />
     <Compile Include="ViewModels\Settings\DesignTimeWindowsSettingsViewModel.cs" />
