+++ Src/MoneyFox.Uwp/Services/NavigationServiceInitializerService.cs
@@ -3,6 +3,8 @@ using MoneyFox.Ui.Shared.ViewModels.Backup;
 using MoneyFox.Ui.Shared.ViewModels.Settings;
 using MoneyFox.Ui.Shared.ViewModels.Statistics;
 using MoneyFox.Uwp.ViewModels;
+using MoneyFox.Uwp.ViewModels.Accounts;
+using MoneyFox.Uwp.ViewModels.Categories;
 using MoneyFox.Uwp.ViewModels.Payments;
 using MoneyFox.Uwp.ViewModels.Settings;
 using MoneyFox.Uwp.ViewModels.Statistic;
