+++ Src/MoneyFox.Uwp/ViewModels/Accounts/EditAccountViewModel.cs
@@ -15,7 +15,7 @@ using System.Globalization;
 using System.Threading.Tasks;
 
 #nullable enable
-namespace MoneyFox.Uwp.ViewModels
+namespace MoneyFox.Uwp.ViewModels.Accounts
 {
     public class EditAccountViewModel : ModifyAccountViewModel
     {
