+++ Src/MoneyFox.Uwp/ViewModels/Accounts/ModifyAccountViewModel.cs
@@ -10,7 +10,7 @@ using System.Globalization;
 using System.Threading.Tasks;
 
 #nullable enable
-namespace MoneyFox.Uwp.ViewModels
+namespace MoneyFox.Uwp.ViewModels.Accounts
 {
     public abstract class ModifyAccountViewModel : ViewModelBase
     {
