+++ Src/MoneyFox.Uwp/ViewModels/Categories/AddCategoryViewModel.cs
@@ -8,7 +8,7 @@ using MoneyFox.Uwp.Services;
 using System.Threading.Tasks;
 
 #nullable enable
-namespace MoneyFox.Uwp.ViewModels
+namespace MoneyFox.Uwp.ViewModels.Categories
 {
     public class AddCategoryViewModel : ModifyCategoryViewModel
     {
