+++ Src/MoneyFox.Uwp/ViewModels/Categories/ISelectCategoryListViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using MoneyFox.Ui.Shared.ViewModels.Categories;
 
 #nullable enable
-namespace MoneyFox.Uwp.ViewModels
+namespace MoneyFox.Uwp.ViewModels.Categories
 {
     /// <summary>
     /// Represents the SelectCategoryListView
