+++ Src/MoneyFox.Uwp/Views/Accounts/AccountListView.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using MoneyFox.Application.Common;
 using MoneyFox.Ui.Shared.ViewModels.Accounts;
-using MoneyFox.Uwp.ViewModels;
+using MoneyFox.Uwp.ViewModels.Accounts;
 using System;
 using Windows.ApplicationModel;
 using Windows.UI.Xaml;
