+++ Src/MoneyFox.Uwp/Views/Accounts/EditAccountView.xaml.cs
@@ -1,4 +1,4 @@
-﻿using MoneyFox.Uwp.ViewModels;
+﻿using MoneyFox.Uwp.ViewModels.Accounts;
 using Windows.UI.Xaml.Controls;
 
 #nullable enable
