+++ Src/MoneyFox.Uwp/Views/Categories/CategoryListControl.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using MoneyFox.Application.Common;
 using MoneyFox.Ui.Shared.ViewModels.Categories;
-using MoneyFox.Uwp.ViewModels;
+using MoneyFox.Uwp.ViewModels.Categories;
 using System;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
