+++ Src/MoneyFox.Uwp/Views/Payments/SelectCategoryDialog.xaml.cs
@@ -1,4 +1,4 @@
-﻿using MoneyFox.Uwp.ViewModels;
+﻿using MoneyFox.Uwp.ViewModels.Categories;
 using Windows.UI.Xaml.Controls;
 
 #nullable enable
